Move elasticsearch api client to a separate package
